add 博客添加、查询 add 图片上传 OSS add 收藏、点赞博客 add 发布博客、删除博客 add 查询博客点赞、收藏量 add 查看用户粉丝数、博客数 add 是否关注、收藏、点赞 add 查询文章浏览量，增加用户浏览记录 add 查询作者文章数，粉丝数，是否关注 add 取消点赞、收藏、取关 add 更新作者总浏览量

+++ src/main/java/com/hjj/homieMatching/model/domain/Blog.java
@@ -54,22 +54,22 @@ public class Blog implements Serializable {
     /**
      * 浏览数
      */
-    private Integer viewNum;
+    private Long viewNum;
 
     /**
      * 点赞数
      */
-    private Integer likeNum;
+    private Long likeNum;
 
     /**
      * 收藏数
      */
-    private Integer starNum;
+    private Long starNum;
 
     /**
      * 评论数
      */
-    private Integer commentNum;
+    private Long commentNum;
 
     /**
      * 创建时间
